+++ src/libqgit2/tests/Repository.cpp
@@ -23,6 +23,9 @@
 #include "qgitrepository.h"
 #include "qgitremote.h"
 
+#include <QPointer>
+#include <QDir>
+
 using namespace LibQGit2;
 
 class TestRepository : public TestBase
@@ -34,55 +37,104 @@ public:
     {}
 
 private slots:
+    virtual void init();
+    virtual void cleanup();
+
     void testRemoteUrlChanging();
     void testLookingUpRevision();
     void testCreateBranch();
     void testDeleteBranch();
+    void testShouldIgnore();
 
 private:
-    Repository repo;
     const QString branchName;
+    QPointer<Repository> repo;
 };
 
+void TestRepository::init()
+{
+    TestBase::init();
+
+    QVERIFY(!repo);
+    repo = new Repository;
+    QVERIFY(repo);
+}
+
+void TestRepository::cleanup()
+{
+    QVERIFY(repo);
+    delete repo;
+    QVERIFY(!repo);
+
+    TestBase::cleanup();
+}
 
 void TestRepository::testRemoteUrlChanging()
 {
-    repo.init(testdir);
+    repo->init(testdir);
 
     const QString remoteName("origin");
-    repo.remoteAdd(remoteName, HttpRemoteUrl);
-    repo.remoteAdd(remoteName, GitRemoteUrl, true);
+    repo->remoteAdd(remoteName, HttpRemoteUrl);
+    repo->remoteAdd(remoteName, GitRemoteUrl, true);
 
-    QScopedPointer<Remote> remote(repo.remote(remoteName));
+    QScopedPointer<Remote> remote(repo->remote(remoteName));
     QCOMPARE(remote->url(), GitRemoteUrl);
 }
 
 void TestRepository::testLookingUpRevision()
 {
-    repo.open(ExistingRepository);
+    repo->open(ExistingRepository);
 
-    Object object = repo.lookupRevision("HEAD^{tree}");
+    Object object = repo->lookupRevision("HEAD^{tree}");
     QCOMPARE(Object::TreeType, object.type());
 }
 
 void TestRepository::testCreateBranch()
 {
     initTestRepo();
-    repo.open(testdir);
+    repo->open(testdir);
 
-    OId head = repo.head().target();
-    QCOMPARE(repo.createBranch(branchName).target(), head);
-    QCOMPARE(repo.lookupShorthandRef(branchName).target(), head);
+    OId head = repo->head().target();
+    QCOMPARE(repo->createBranch(branchName).target(), head);
+    QCOMPARE(repo->lookupShorthandRef(branchName).target(), head);
 }
 
 void TestRepository::testDeleteBranch()
 {
     initTestRepo();
-    repo.open(testdir);
-    repo.createBranch(branchName);
+    repo->open(testdir);
+    repo->createBranch(branchName);
 
-    repo.deleteBranch(branchName);
-    EXPECT_THROW(repo.lookupShorthandRef(branchName), Exception);
+    repo->deleteBranch(branchName);
+    EXPECT_THROW(repo->lookupShorthandRef(branchName), Exception);
+}
+
+void TestRepository::testShouldIgnore()
+{
+    initTestRepo();
+    repo->open(testdir);
+
+    const QString ignoredFileName("Makefile");
+    const QString includedFileName("notignored.txt");
+
+    // Relative paths
+    QVERIFY(repo->shouldIgnore(ignoredFileName));
+    QVERIFY(!repo->shouldIgnore(includedFileName));
+
+    QDir dir(testdir);
+    QString testDirName(dir.dirName());
+
+    // Absolute paths
+    QVERIFY(repo->shouldIgnore(dir.absoluteFilePath(ignoredFileName)));
+    QVERIFY(!repo->shouldIgnore(dir.absoluteFilePath(includedFileName)));
+
+    dir.cdUp();
+
+    // Path containing .. but still leading up to the repository
+    QVERIFY(!repo->shouldIgnore(dir.absoluteFilePath(dir.absolutePath() + "/../" + dir.dirName() + "/" + testDirName + "/" + includedFileName)));
+
+    // Absolute path outside the repository
+    EXPECT_THROW(repo->shouldIgnore(dir.absoluteFilePath(ignoredFileName)), Exception);
 }
 
 QTEST_MAIN(TestRepository)